[#WEB-141] Add group management - update

+++ dcm4chee/dcm4chee-usr/trunk/dcm4chee-usr-dao/src/main/java/org/dcm4chee/usr/dao/UserAccessBean.java
@@ -51,7 +51,15 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.UUID;
 
+import javax.annotation.PostConstruct;
 import javax.ejb.Stateless;
+import javax.management.AttributeNotFoundException;
+import javax.management.InstanceNotFoundException;
+import javax.management.MBeanException;
+import javax.management.MBeanServer;
+import javax.management.MalformedObjectNameException;
+import javax.management.ObjectName;
+import javax.management.ReflectionException;
 import javax.persistence.EntityManager;
 import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
@@ -61,7 +69,9 @@ import net.sf.json.JSONObject;
 
 import org.dcm4chee.usr.entity.User;
 import org.dcm4chee.usr.entity.UserRoleAssignment;
+import org.dcm4chee.usr.sar.UserAccessService;
 import org.jboss.annotation.ejb.LocalBinding;
+import org.jboss.mx.util.MBeanServerLocator;
 import org.jboss.system.ServiceMBeanSupport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -73,21 +83,86 @@ import org.slf4j.LoggerFactory;
  */
 @Stateless
 @LocalBinding(jndiBinding=UserAccess.JNDI_NAME)
-public class UserAccessBean extends ServiceMBeanSupport implements UserAccess {
+public class UserAccessBean implements UserAccess {
 
     private static Logger log = LoggerFactory.getLogger(UserAccessBean.class);
     
     @PersistenceContext(unitName="dcm4chee-usr")
     private EntityManager em;
-    
-    private String roleFilename = "roles.json";
 
-    public void setRoleFilename(String roleFilename) {
-        this.roleFilename = roleFilename;
+    private ObjectName serviceObjectName;
+    private final MBeanServer server = MBeanServerLocator.locate();
+
+    private String userRoleName;
+    private String adminRoleName;
+
+    public void init(String serviceObjectName) {
+
+        BufferedReader reader = null;
+        try {
+            this.serviceObjectName = new ObjectName(serviceObjectName);
+            userRoleName = 
+                (String) this.server.getAttribute(
+                    this.serviceObjectName, 
+                    "userRoleName");
+
+            adminRoleName = 
+                (String) this.server.getAttribute(
+                    this.serviceObjectName, 
+                    "adminRoleName");
+        
+            List<Role> roleList = new ArrayList<Role>();
+
+            if (new File(this.getRoleFilename()).exists()) {
+                String line;
+                reader = new BufferedReader(new FileReader(this.getRoleFilename()));
+                while ((line = reader.readLine()) != null)
+                    roleList.add((Role) JSONObject.toBean(JSONObject.fromObject(line), Role.class));
+
+                if (roleList.size() > 0) {
+                    boolean haveUserRoleName = false;
+                    boolean haveAdminRoleName = false;
+                    
+                    for (Role role : roleList) {
+                        if (!role.isSystemRole()) break;                       
+                        if (role.getRolename().equals(userRoleName))
+                            haveUserRoleName = true;
+                        if (role.getRolename().equals(adminRoleName))
+                            haveAdminRoleName = true;
+                    }
+                    if (!haveUserRoleName) addRole(new Role(userRoleName, true));
+                    if (!haveAdminRoleName) addRole(new Role(adminRoleName, true));
+                    return;
+                }
+            }
+            addRole(new Role(userRoleName, true));
+            addRole(new Role(adminRoleName, true));
+        } catch (Exception e) {
+            log.debug("Exception: ", e);
+        } finally {
+            try {
+                reader.close();
+            } catch (Exception ignore) {}
+        }           
+    }
+
+    public String getUserRoleName() {
+        return this.userRoleName;
+    }
+
+    public String getAdminRoleName() {
+        return this.adminRoleName;
     }
 
     public String getRoleFilename() {
-        return roleFilename;
+        try {
+            return (String) this.server.getAttribute(
+                    this.serviceObjectName, 
+                    "roleFilename");
+        } catch (Exception e) {
+            log.debug("Exception: ", e);
+        }
+        return null;
     }
 
     public User getUser(String userId) {
@@ -152,45 +227,7 @@ public class UserAccessBean extends ServiceMBeanSupport implements UserAccess {
         .setParameter("rolename", ura.getRole())
         .executeUpdate();
     }
-
-    public void checkSystemRoles(String userRoleName, String adminRoleName) {
-        List<Role> roleList = new ArrayList<Role>();
-        BufferedReader reader = null;
-        try {
-            if (new File(this.getRoleFilename()).exists()) {
-                String line;
-                reader = new BufferedReader(new FileReader(this.getRoleFilename()));
-                while ((line = reader.readLine()) != null)
-                    roleList.add((Role) JSONObject.toBean(JSONObject.fromObject(line), Role.class));
-
-                if (roleList.size() > 0) {
-                    boolean haveUserRoleName = false;
-                    boolean haveAdminRoleName = false;
-                    
-                    for (Role role : roleList) {
-                        if (!role.isSystemRole()) break;                       
-                        if (role.getRolename().equals(userRoleName))
-                            haveUserRoleName = true;
-                        if (role.getRolename().equals(adminRoleName))
-                            haveAdminRoleName = true;
-                    }
-                    if (!haveUserRoleName) addRole(new Role(userRoleName, true));
-                    if (!haveAdminRoleName) addRole(new Role(adminRoleName, true));
-                    return;
-                }
-            }
-            addRole(new Role(userRoleName, true));
-            addRole(new Role(adminRoleName, true));
-        } catch (FileNotFoundException e) {
-            log.debug("Exception: ", e);
-        } catch (IOException e) {
-            log.debug("Exception: ", e);
-        } finally {
-            try {
-                reader.close();
-            } catch (Exception ignore) {}
-        }           
-    }
+    
     
     public List<Role> getAllRolenames() {
         List<Role> roleList = new ArrayList<Role>();
@@ -237,6 +274,7 @@ public class UserAccessBean extends ServiceMBeanSupport implements UserAccess {
     public void addRole(Role role) {
         BufferedWriter writer = null; 
         try {
+            if (roleExists(role.getRolename())) return;
             writer = new BufferedWriter(new FileWriter(this.getRoleFilename(), true));
             JSONObject jsonObject = JSONObject.fromObject(role);
             if (role.getRolename() == null) jsonObject.put("rolename", JSONNull.getInstance());
